+++ raiden/api/python.py
@@ -868,6 +868,7 @@ class RaidenAPI:  # pragma: no unittest
         transfer_timeout: int = None,
         secret: Secret = None,
         secrethash: SecretHash = None,
+        locktimeout: BlockTimeout = None,
     ):
         """ Do a transfer with `target` with the given `amount` of `token_address`. """
         # pylint: disable=too-many-arguments
@@ -880,6 +881,7 @@ class RaidenAPI:  # pragma: no unittest
             identifier=identifier,
             secret=secret,
             secrethash=secrethash,
+            locktimeout=locktimeout,
         )
         payment_status.payment_done.wait(timeout=transfer_timeout)
         return payment_status
@@ -893,6 +895,7 @@ class RaidenAPI:  # pragma: no unittest
         identifier: PaymentID = None,
         secret: Secret = None,
         secrethash: SecretHash = None,
+        locktimeout: BlockTimeout = None,
     ):
         current_state = views.state_from_raiden(self.raiden)
         token_network_registry_address = self.raiden.default_registry.address
@@ -958,6 +961,7 @@ class RaidenAPI:  # pragma: no unittest
             identifier=identifier,
             secret=secret,
             secrethash=secrethash,
+            locktimeout=locktimeout,
         )
         return payment_status