+++ raiden/storage/serialization/types.py
@@ -60,6 +60,31 @@ from raiden.utils.typing import (
     WithdrawAmount,
 )
 
+MESSAGE_NAME_TO_QUALIFIED_NAME = {
+    "AuthenticatedMessage": "raiden.messages.abstract.AuthenticatedMessage",
+    "Delivered": "raiden.messages.synchronization.Delivered",
+    "EnvelopeMessage": "raiden.messages.transfers.EnvelopeMessage",
+    "LockedTransferBase": "raiden.messages.transfers.LockedTransferBase",
+    "LockedTransfer": "raiden.messages.transfers.LockedTransfer",
+    "LockExpired": "raiden.messages.transfers.LockExpired",
+    "PFSCapacityUpdate": "raiden.messages.path_finding_service.PFSCapacityUpdate",
+    "PFSFeeUpdate": "raiden.messages.path_finding_service.PFSFeeUpdate",
+    "Ping": "raiden.messages.healthcheck.Ping",
+    "Pong": "raiden.messages.healthcheck.Pong",
+    "Processed": "raiden.messages.synchronization.Processed",
+    "RefundTransfer": "raiden.messages.transfers.RefundTransfer",
+    "RequestMonitoring": "raiden.messages.monitoring_service.RequestMonitoring",
+    "RevealSecret": "raiden.messages.transfers.RevealSecret",
+    "SecretRequest": "raiden.messages.transfers.SecretRequest",
+    "SignedMessage": "raiden.messages.abstract.SignedMessage",
+    "SignedRetrieableMessage": "raiden.messages.abstract.SignedRetrieableMessage",
+    "ToDevice": "raiden.messages.matrix.ToDevice",
+    "Unlock": "raiden.messages.transfers.Unlock",
+    "WithdrawConfirmation": "raiden.messages.withdraw.WithdrawConfirmation",
+    "WithdrawExpired": "raiden.messages.withdraw.WithdrawExpired",
+    "WithdrawRequest": "raiden.messages.withdraw.WithdrawRequest",
+}
+
 
 def transfer_task_schema_serialization(task: TransferTask, parent: Any) -> Schema:
     # pylint: disable=unused-argument